# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Highleads API",
        "description": "API highleads-backend",
        "contact": {
            "name": "Marco Aurélio",
            "email": "marcoascabral@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/blacklist/{id}": {
            "get": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "Exibe um registro da blacklist",
                "description": "Exibe um registro da blacklist por seu ID.",
                "operationId": "d5ab44f3d139eafb76d2612de7c4835f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da blacklist",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados da blacklist."
                    },
                    "404": {
                        "description": "Item não encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "Alterar um registro da blacklist",
                "description": "Alterar um registro da blacklist",
                "operationId": "21a49eb0f37fb96f7dfd8a27a06cacaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da entrada na blacklist",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Dados do registro da blacklist",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "texto"
                                ],
                                "properties": {
                                    "texto": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "https://zoit.com.br"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Salvo com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao inserir."
                    },
                    "404": {
                        "description": "Não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "Exclui uma entrada da blacklist",
                "description": "Exclui a entrada na blacklist por seu ID.",
                "operationId": "4f9fc23f5a723b73c3b110631aa90485",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Excluído com suceso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "excluído": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/blacklist-search?email{email}": {
            "get": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "Procurar o item da blacklist por email",
                "description": "Procurar o item da blacklist por email",
                "operationId": "47a0c1d58f6d29a78d2799251be358c4",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste@dominio.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados da blacklist."
                    }
                }
            }
        },
        "/v1/blacklist": {
            "post": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "Criar uma nova entrada na blacklist",
                "description": "Criar uma nova entrada na blacklist",
                "operationId": "ea418b301e3ad450fc0f7b1996c938e7",
                "requestBody": {
                    "description": "Dados da blacklista",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "texto"
                                ],
                                "properties": {
                                    "texto": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blacklist salva com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "created": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao inserir o registro."
                    }
                }
            }
        },
...
    }
}